trunk: changeset 419
Client windows are unmapped when frames are unmapped. This was an
overlooked (but lame and redundant) requirement of the ICCCM and
fullfilling it might fix some apps. (Ion probably still is far from
ICCCM-compliant, but so are most of the badly behaving apps.)

darcs-hash:20030411160602-e481e-3944ff763f1f6c500d591662368c9ac3b68648e8.gz