completefile-groupexec
Although ion-completefile does check if any matching executable is
world-executable or user-executable, it doesn't check for
group-executable. This is problematic for example on systems, where
games are installed with r-xr-x--- and root:games (which seems to be a
common method). This introduces a check for the group id, basically it
gets all groups for the current user and checks if any of them is
applicable. This is only done when the user and world checks already
have failed and the group-executable bit is set.

darcs-hash:20060211153509-e8521-4e88feafae35dc8ab110bd578f67032c5980b983.gz