Use clock_gettime(CLOCK_MONOTONIC) when available.
*nix provides only one interval timer (or 'select' timeout), so programs
with multiple timers need to do their own timeout management. This means
converting the times from delays to absolute times. When gettimeofday is
used, the time thus provided is not guaranteed to be monotonic, and may
be adjusted backwards, which is a serious problem. The CLOCK_MONOTONIC
time advanced monotonically from an unspecified starting point, thus
providing exactly what is needed for decent timeout management.

This partially fixes issues with e.g. statusbar updates when time is
adjusted backwards. If CLOCK_MONOTONIC is not available, the issue
remains, as I can't be arsed to write ugly countermeasures against
the lack of an absolutely _essential_ feature that every system that
is not totally moronic, should have had to begin with. (I should also
be just using the rather nice libevent that I just learned about
instead of my own mainloop library... but since there's a feature
freeze and the conversion would be quite a bit of work that could
induce new bugs, it isn't happening for Ion3.)

darcs-hash:20070524205718-4eba8-f6a0c47c8da5f0d13e414163d59534cb5d97da44.gz