More sophisticated pointer event hack.
Apparently the old one was needed, after all, for actions on transients.
So now we do like this: if the subwindow from the event listens to the
button/modifier combination in question at all, then we let it handle it,
blocking the parent. Otherwise the parent handles it.

darcs-hash:20071201174632-4eba8-ef79329f28689c39b6e4b02888e8e54a253a06e7.gz