Handle XConfigureRequestEvent's moving transient dialogs to incorrect locations
Some applications, in particular Java apps, tend to move transient dialogs to
invalid locations by way of XConfigureRequestEvent's.

This patch ignores X or Y values from XConfigureRequestEvent that are outside
the bounds of the frame.

When X/Y values are not out-of-bounds but still close to the right/bottom edge,
the dialog might get 'squashed'. To prevent this, we now consider width and
height to be non-'weak' even if only new X or Y values are submitted.

https://sourceforge.net/tracker/?func=detail&aid=3167262&group_id=314802&atid=1324528