Do not include 'features.h' to make use feature test macro's
The feature test macros (such as _POSIX_C_SOURCE) are somewhat standard/
portable. 'features.h' is a Linux/glibc thing and should not be included
directly. It is automatically included by other header files (in this case
<sys/select.h>) as required. This is documented in the NOTES section of the
feature_test_macros manpage on linux. Verified with Silas Sila that this
also indeed behaves as expected (uses pselect) on NetBSD.