EWMH compliance fix: _NET_WM_NAME needs to be of type UTF8_STRING
In case Xlib doesn't support XUTF8StringStyle we fall back to the old, broken
behaviour of using the encoding of the current locale. If there's really
anyone still using such a beast _and_ caring about EWMH [1] they can enhance
the code to do an explicit conversion (locale enconding to UTF-8) and set the
type to the UTF8_STRING Atom manually.

[1] http://freedesktop.org/wiki/Specifications/wm-spec

Signed-off-by: Sascha Silbe <sascha-pgp@silbe.org>