Remove call to lua_cpcall for Lua 5.2
In Lua 5.2 lua_cpcall was removed. The manual[1] suggests to replace it
by an appropriate call to lua_pcall.

TODO: Do we need to make a call to lua_checkstack before pushing the
function and its arguments to be safe here? If so, what's an appropriate
argument to lua_checkstack?

[1] http://www.lua.org/manual/5.2/manual.html#8.3