Mark all frames involved in a maximization as maximized
While this is of course not technically accurate, is has the effect that
we undo a maximization whenever we toggle maximization on a frame that
is already part of a maximized split. This is desirable for two reasons:

First, it gives the expected result on frames that have optically been
maximized in a maximization process but were not the one that we toggled
maximization on:
-----               -----
|   | vert. max. 1  | | |
|---| ------------> |1|2|
|1|2|               | | |
-----               -----
Before, trying to untoggle maximization on frame 2 would instead try to
maximize frame 2 which is undesirable. With the new behavior a user does
not have to remember on which frame he toggled maximization to undo a
maximization. This was explicitly requested in the discussion with the
subject "Toggle horizontal/vertical maximize behaviour" on ion-general.

Secondly, it even seems desirable in the following situation, where a
frame has not been optically maximized.
-----               -----
| 0 | vert. max. 3  | | |
|---| ------------> |1| |
|1| |               |-|3|
|-|3|               |2| |
|2| |               | | |
-----               -----
One might argue that restoring the layout when toggling vertical
maximization on 2 is wrong because it is not maximized yet, but
maximizing frame 2 in this setup would have the ugly side effect of
forever losing the original layout and turning frame 0 into an
unresponsive frame of height 0. To achieve the desired result a user
will now have to toggle maximization on 2 twice, first to restore the
actual layout and then to actually maximize it. I find this acceptable,
given that the end result is much cleaner.

On a side node, this commit also tries to avoid any flickering for the
statusbar or the dock.