When a screen is detached, move its workspaces to the first screen
This is better than the old behavior (where these workspaces simply became
unavailable).

This needs further improvement: for example, once a screen is re-attached, it
would be good to move workspaces back to it. Thus, we need an algorithm to find
out which monitor each workspace prefers to live on.