Don't ignore the stdisp after all, use less precise verification
Removing and readding the stdisp had ugly side effects, like the
flickering I've been trying to address with several commits already. New
bugs kept showing up and the approach was getting more and more
complicated when I tried to fix them. So I've decided to not ignore the
stdisp after all. This means that it doesn't make sense to save the
geometry for all splits, because the statusbar moves around on
horizontal maximization, introducing new splits and messing with the
existing ones. I now use a very naive form of verification before
restoring the layout, which only consists in checking that all frames
have the FRAME_MAXED_* flag set, compare the last commit. This makes
things a lot easier and it should be precise enough.

TODO: The statusbar is not restored when undoing a horizontal
maximization. Two things are causing problems:
1) Simply calling a tiling_manage_stdisp or similar at the end of the
   restore process causes inacceptable flickering.
2) Resizing the statusbar will cause other windows to be resized. This
   will possibly lose a FRAME_MAXED_VERT flag on a frame.
Right now the statusbar will be automatically restored when for example
switching away from and back to the current workspace. Under the right
circumstances this causes problem 2) above.