Handle case of non-fs, horiz. stdisp in bl corner
This seems to take care of the two problems mentioned in the message to
commit 8370a156ae6440e3983486ca4c7a167caf72cf25 in the special case that
the potential non-fullsized stdisp is positioned in the lower left
corner and has a horizontal orientation. This should be the most
commmon case. Adding support for other cases is not difficult (maybe a
horizontal stdisp in the upper left corner is not that uncommon?), but
it will only happen on request.