Change handling of geometry of stdisp
Before we set the geometry of the stdisp to its correct target geometry.
This however renders the whole tiling inconsistent. Also taking into
account the geometry changes from mmoving away from the stdisp renders
the tiling inconsistent. We tried to counter this by calling
tiling_manage_stdisp, but it turns out that this is not good enough to
repair the tiling.

So now we resize the stdisp to the geometry that is necessary to keep
the tiling consistent and then call split_sanitize_stdisp. Also we only
use the geometry calculated for frames moving away from the stdisp for
updating the window informations, while using a fake geometry
information for the split without these changes. split_sanitize_stdisp
will then update the tiling in a way that again accomodates the frames
that are temporarily bigger than their surrounding splits. Still should
combine the advantages of using the correct geometry (which massively
avoids flickering in programs with resizing problems, notably ncurses
based ones) and of having consistent tiling (desirable for obvious
reasons).