Replace deprecated XKeycodeToKeysym with XkbKeycodeToKeysym
Thanks Philipp for noticing. While XKB is strictly spoken an extension, it
looks like it has been available through the standard X libraries on all major
systems (linux, freebsd, cygwin) for quite some time now. Not entirely sure
about Solaris. We can make this a system.mk option (default on) if this is a
problem for any current system.