Update copyright header of de/fontset.c
Most of the code seems to be a verbatim copy of
<http://www.debian.org/doc/manuals/intro-i18n/ch-examples.en.html#s13.4.5>
which is licensed under the GPL, so that we can hardly ship it with an
MIT license copyright header. I don't see any similarity whatsoever to
the the file "Screen.cc for Blackbox - an X11 Window manager" mentioned
in the header (which is indeed published under the MIT license). So I've
simply dropped the reference to the latter file and put this file under
the GPL.