add notes about getting and updating the source to README
I struggled a bit to update to the latest version of notion because
new submodules were added.  The natural place to look first was the
README file, which didn't tell anything about this problem, except for
the wiki link.  I think it would be nice to have a note about this in
the README.