Use META+(Shift+)scrollwheel to switch windows
Works nicely for the common cases where the mouse pointer is in the focused
window. When the mouse pointer is not in the focussed window, I'm not entirely
sure what behavior is desirable. Right now it scrolls through the tabs of the
frame with the mouse pointer for horizontal scrolling, and behaves somewhat
erratically for vertical scrolling.