Only allow removal of last workspace if there are multiple screens
This is a more conservative version of
ca7765688c5532f55281089a66a42b176f7849e5: it preserves the behavior of
Ion3 in case there is a single screen.

But we still want to be able to remove the last workspace from a screen
if there are multiple screens: this is used in the redistribution of
workspaces on detaching and attaching screens.