Also support charge_* files with sysfs
The current implementation does not account for cases where there is an
energy_now file and no energy_full file, but a charge_full file*. This
is because I doubt that such systems exist.

To also include this case, one would have to convert between charge and
energy values using the information in the voltage_now file.

* or the other way around