Be more forgiving when XCreateFontSet creates a fontset with missing charsets
Especially utf-8 locales list unreasonable amounts of charsets as requirements.
When a font can be found that satisfies some, but not all of them, we just use
the font and write a warning into the logging. A user who really cares about
all those charsets can fix his configuration himself.