focuslist is now populated after a delay
The focus list is no longer immediately updated. This is nice because you can
now search for a window by quickly cycling through some windows without them
showing up near the top of the focus list. This allows the
go-to-previous-windows functions to keep working as the user intends. This
behavior is enabled with a config option: 'focuslist_insert_delay'. This is
turned off by default.

Before this patch, the currently-focused region was assumed to be the one at the
top of the focuslist. This is no longer true 100% of the time, so those two
things are now tracked separately