build: all .o objects no longer depend on exports.h
The build system had an unconditional dependency declared:

 %.o: %.c $(EXPORTS_H)

The result of this was that when any .c file was changed, lua would regenerate
its exports, which would then trigger a rebuild of EVERYTHING. In truth, most
things don't really depend on exports.h.

This patch explicitly declares dependencies on exports.h. The dependency
generator generates these dependencies too, but the explicit declarations are
required for the initial build, when the generated dependencies aren't yet
available