build: dependencies are now ALWAYS generated. Also simpler implementation
Previously you didn't get header dependencies, unless you asked for them. Now
this information is always generated, and it's done together with the
compilation step. This is simple and robust. The only way a dependency
relationship is missed is the first time you add a #include. Rebuilding that
source file once adds the dependency