Fixed workspace-indicator issue triggered by IntelliJ IDEA
Normally the workspace_indicator_timeout config option creates an indicator with
the name of the current workspace in the corder of the screen. I was observing
an issue where switching workspaces away from an IntelliJ IDEA window causes the
indicator to not be displayed. This was due to logic in notion that removes the
indicator when an NULL or unnamed screen or mplex are focussed. Removing this
logic fixes the issue, and I can find no reason this was needed in the first
place. Things work well without.