Experimenting with combining submap and kpress (for #93)
Experimenting with allowing 'submap' and 'kpress' handlers to the
same key combination.

Unfortunately when throwing the submaps and the kpress handlers on one heap
the kpress handler is often not found, since a submap without kpress handler
is found earlier in the 'hierarchy'. This is not so much of a problem otherwise
because kpress handlers are less likely to conflict among each other.

I think we should introduce a new 'action' for submaps to separate the two.