Support for showing tab numbers while a key is pressed (#103)
With this change, you can register keybindings to show tab numbers while
holding a key, e.g.:

```
defbindings("WFrame.toplevel", {
    kpress("Super_L", "WFrame.set_numbers(_, 'during_grab')"),
})
```

The grab is necessary to be able to restore the normal titles again when the
key is released. A feature was added to 'grab' to allow grabbing the keyboard
while still processing mouse events (such as resizing frames with META+Button3).