Improved Lua version detection and Lua 5.3 support (#132)
* libtu/Makefile: it's C89, not C98

* Lua 5.3 removed deprecated variadic function `arg' parameter

Variadic functions were introduced in 5.0, but the method of accessing
them changed in 5.1. This makes the code incompatible with 5.0 (it is
already incompatible since at least b2b17559d), which will be removed
from the build system in a subsequent commit.

Fixes #124. See also #90.

* Remove Lua 5.0 from the build system and improve lua-detect.mk

Lua doesn't include an official pkg-config file, therefore we can't rely on
pkg-config variables like "V" or "major_version" - on my system, it detected
the 5.3 library/cflags and used the 5.1 lua/luac binaries (I have both installed
and there was a fallback to 5.1 if it couldn't query these variables from
pkg-config).