Allow creating the statusbar while brushes aren't available yet #233
Because loading the look was moved to after loading the rest of the config,
so that you can use the modules (e.g. for determining the screen size)
while building the look.

If this turns out to be too harsh, we could separate the loading of the
drawing engine from the loading of the look itself.