Relax module version check strictness (fixes #101)
Given a version number such as '4.0.2-5-g4d284b74+20210309-1735', prior
to this commit notion would refuse to load modules that don't match the
exact string. Now it only checks up until the first minus and prints a
warning in case of a mismatch, which still pops up in a nasty window so
the user knows about it but doesn't temporarily break functionality.